Refactor components to common folder
